API

Exported members

Core types

Results.Result — Type

Synonym for Union{Ok{T}, Err{E}}.

As well as working with the Result combinators defined below, Results implement the following protocols:

iterate: Yields one T if the Result is Ok, yields nothing otherwise.

length: Returns 1 if the Result is Ok, 0 if Err

Results.Option — Type

Synonym for Union{Some{T}, None}.

Options implement the following protocols:

iterate: Yields one T if the Option is Some, yields nothing otherwise.

length: Returns 1 if the Result is Ok, 0 if Err

Results.Ok — Type

Represents an Ok result of computation.

Results.Err — Type

Represents an computation error.

Results.None — Type

Shorthand for Nothing.

Results.none — Constant

Shorthand for nothing.

Input functions

Results.to_option — Function

to_option(value::Union{T, Nothing})::Option{T}
to_option(value::Result)::Option

Converts a nullable value to an Option. Also converts a Result to an Option, turning Ok into Some and Err into None.

Examples

julia> to_option("value")
Some("value")
julia> to_option(Some(5))
Some(Some(5))
julia> to_option(nothing) |> println
nothing

Results.to_result — Function

to_result(value::T, err)::Ok{T}
to_result(value::None, err::E)::Err{E}
to_result(value::None, err::Function)::Err

Converts a nullable value to a Result, using the supplied error value in place of a nothing. Also converts an Option into a Result, turning Some into Ok and None into Err.

Examples

julia> to_result("value", "error")
Ok("value")
julia> to_result(nothing, "error")
Err("error")
julia> to_result("value", () -> println("lazily calculates errors"))
Ok("value")
julia> to_result(Ok(5), nothing)
Ok(Ok(5))

Output functions

Results.unwrap — Function

unwrap(o::Option{T})::T
unwrap(r::Result{T, E})::T
unwrap(o::Union{Option, Result}, error::Function)
unwrap(o::Union{Option, Result}, error::Exception)
unwrap(o::Union{Option, Result}, error::String)

Unwrap an Ok value. Throws an error if r is Err instead.

In the two argument form, error is raised if it is an Exception. If it is a string, it is passed as a message to UnwrapError. If it is a function, it is called with an error value to produce an error.

Examples

julia> unwrap(Ok(5))
5
julia> unwrap(Err(0))
ERROR: Results.UnwrapError("unwrap() called on an Err: 0")
julia> unwrap(none)
ERROR: Results.UnwrapError("unwrap() called on None")
julia> unwrap(none, "value is none")
ERROR: Results.UnwrapError("value is none")
julia> unwrap(nothing, BoundsError([1,2]))
ERROR: BoundsError: attempt to access 2-element Array{Int64,1}
julia> unwrap(Err(5), v -> "Error value '" * string(v) * "'")
ERROR: Results.UnwrapError("Error value '5'")

Results.unwrap_or — Function

unwrap_or(o::Option{T}, default)::T
unwrap_or(r::Result{T, E}, default)::T

Unwrap an Ok value, or return default. default may be T or a function returning T.

Examples

julia> unwrap_or(Ok("value"), "error")
"value"
julia> unwrap_or(Err(5), "error")
"error"
julia> unwrap_or(Some("value"), () -> begin println("Generating error"); "error" end)
"value"
julia> unwrap_or(None(), () -> begin println("Generating error"); "error" end)
Generating error
"error"

Results.to_nullable — Function

to_nullable(val::Option{T})::Union{T, Nothing}
to_nullable(val::Result{T, E})::Union{T, Nothing}

Convert a Result{T, E} or Option{T} into a simple nullable value Union{T, Nothing}. Note that this may lose structure (this transformation is non-injective, as Some(nothing) and nothing map to the same value).

Combinators

Results.try_map — Function

try_map(f, opt::Option...)::Option
try_map(f, rslt::Result...)::Result
try_map(f::Option, rslt::Option...)::Option
try_map(f::Result, rslt::Result...)::Result

Map f over the contents of an Ok or Some value, leaving a None or Err value untouched.

Example

julia> try_map((x) -> 2*x, Ok(5))
Ok(10)
julia> try_map(+, Err("First error"), Err("Second error"))
Err("First error")
julia> try_map.((x) -> 2*x, [Ok(5), Err("missing value")])
2-element Array{Any,1}:
 Ok(10)
 Err("missing value")
julia> try_map(Ok(+), Ok(5), Ok(10))
Ok(15)

Results.map_err — Function

map_err(f, result::Result)::Result

Map f over the contents of an Err value, leaving an Ok value untouched.

Results.and_then — Function

and_then(result::Result, funcs...)::Result

Bind result to the proceeding functions. If result is Ok, its contents will be passed to each function in turn. Any Err value will be returned immediately.

Examples

julia> Ok("Build") ⊗ val -> Ok(string(val, " a ")) ⊗ val -> Ok(string(val, "string"))
Ok("Build a string")
julia> Err("Error") ⊗ val -> Ok(string(val, " a ")) ⊗ val -> Ok(string(val, "string"))
Err("Error")
julia> Ok("Build") ⊗ val -> Err("Error") ⊗ function (val) error("long circuited"); Ok("value") end
Err("Error")

and_then(option::Option, funcs...)::Option

Bind option to the proceeding functions. While option is Some, its contents will be passed to each function in turn. Any None will be returned immediately.

Examples

julia> Some("Build") ⊗ val -> Some(string(val, " a ")) ⊗ val -> Some(string(val, "string"))
Some("Build a string")
julia> none ⊗ val -> Some(string(val, " a ")) ⊗ val -> Some(string(val, "string"))

julia> Some("Build") ⊗ val -> none ⊗ function (val) error("long circuited"); Some("value") end

and_then(func::Base.Callable)::Function

Partially-applied version of and_then.

Results.try_collect — Function

try_collect(iter)
try_collect(result::Result, results...)::Result{Vector}
try_collect(option::Option, options...)::Option{Vector}

Collect an iterator of Options or Results into a single Option or Result containing an array. Short-circuits on error.

Examples

julia> try_collect([Ok(5), Ok(10), Ok(3)])
Ok([5, 10, 3])
julia> try_collect([Ok(10), Err("err1"), Err("err2")])
Err("err1")
julia> try_collect([Ok(10), None(), Ok(5)])  # be careful mixing Result with Option!
Ok(Union{Nothing, Int64}[10, nothing, 5])

Results.flatten — Function

flatten(o::Option{Option{T}})::Option{T}
flatten(r::Result{Result{T, E1}, E2})::Result{T, Union{E1, E2}}

Flatten one layer of a nested Option or Result type.

Examples

julia> Ok(Ok(5)) |> flatten
Ok(5)
julia> Ok(Err("inner")) |> flatten
Err("inner")
julia> Some(Some(Some(5))) |> flatten
Some(Some(5))
julia> Ok(Some(5)) |> flatten  # mixed Option and Result types not supported
ERROR: MethodError

Results.ok — Function

ok(r::Result{T, E})::Option{T}

Convert a Result into an Option, discarding any Err value.

Examples

julia> ok(Ok(5))
Some(5)
julia> ok(Err("error")) |> println
nothing

Results.ok_or — Function

ok_or(o::Some{T}, err)::Ok{T}
ok_or(::None, err::E)::Err{E}
ok_or(::None, err::Function)::Err

Convert an Option into a Result, using the supplied error value in place of None.

Examples

julia> ok_or(Some(5), "error")
Ok(5)
julia> ok_or(None(), "error")
Err("error")
julia> ok_or(None(), () -> "lazy error")
Err("lazy error")
julia> None() |> ok_or("partially applied")
Err("partially applied")

Predicate functions

Results.is_ok — Function

is_ok(r::Result)::Bool

Return whether a Result is Ok or Err.

Results.is_err — Function

is_err(r::Result)::Bool

Return whether a Result is Err.

Results.is_some — Function

is_some(o::Option)::Bool

Return whether an Option is Some.

Results.is_none — Function

is_none(o::Option)::Bool

Return whether an Option is None. Equivalent to Base.isnothing.

Results.has_val — Function

has_val(result::Union{Result, Option})::Bool

Returns true for a successful Result or Option.

Collection utilities

Results.try_pop! — Function

Try to pop a value from a collection.

Results.try_get — Function

Try to get a value from an array or collection.

Results.try_peek — Function

try_peek(iter; state=missing)::Option

Try to get the next value from an iterator. If state is not missing, use it in the call to iterate.

Results.try_first — Function

try_first(c)::Option

Try to get the first element from a collection. Uses Base.firstindex under the hood.

try_first(s::AbstractString, n::Integer)::Option{AbstractString}

Try to get the first n chars from a string.

Results.try_last — Function

try_last(c)::Option

Try to get the last element from a collection.

Macros

Results.@catch_result — Macro

Catches an exception inside expr and returns a Result instead.

If a type is given, only exceptions of that type will be caught.

Examples

julia> @catch_result begin
           arr = [5,3,2]
           arr[4]
       end
Err(BoundsError([5, 3, 2], (4,)))
julia> @catch_result [5,3,2][3]
Ok(2)
julia> @catch_result TypeError [5,3,2][4]
ERROR: BoundsError: attempt to access 3-element Array{Int64,1} at index [4]

Results.@if_let — Macro

Run then_block if the assignment expression returns an Ok or Some value. Runs else_block otherwise.

Example

julia> @if_let val = Some(5) begin
           2*val
       end begin
           0
       end
10
julia> @if_let val = Err("error") begin
           2*val
       end begin
           0
       end
0
julia> @if_let val = Err("error") begin
           println(val)
       end

Results.@some_if — Macro

If predicate, evaluate the enclosed expression wrapped in Some. Otherwise, return None.

Example

julia> try_get(a, index) = @some_if isassigned(a, index) a[index]
try_get (generic function with 1 method)
julia> try_get([2,3,4], 2)
Some(3)
julia> try_get([1,2,3], 10)

Results.@try_unwrap — Macro

Unwraps an Ok or Some value, while returning error values upstream. Highly useful for chaining computations together.

Example

julia> function test(x::Result)::Result
           y = @try_unwrap(x) .- 5
           z = @try_unwrap try_pop!(y) |> ok_or("Empty array")
           Ok(z)
       end
test (generic function with 1 method)

julia> test(Ok([5, 8]))
Ok(3)
julia> test(Ok([]))
Err("Empty array")
julia> test(Err(5))
Err(5)

Results.@unwrap_or — Macro

Macro version of unwrap_or, which allows for the embedding of control statements in the or clause.

Example

julia> for v in [[2,3,4], [3,4,5], [], [1]]
           println(@unwrap_or(try_get(v, 1), break))
       end
2
3

Results.@while_let — Macro

Loop block while the assignment expression assign returns an Ok or Some value.

Example

julia> a = [1,2,3];

julia> @while_let val = try_pop!(a) begin
           print(val)
       end
321

Un-exported members

Results.UnwrapError — Type

Exception thrown when unwrap() is called on an Err

Results.:← — Function

Shorthand for try_map. Enter as \leftarrow.

Example

julia> Ok(x -> 2*x) ← Ok(5)
Ok(10)
julia> x -> 2*x ← Ok(5)  # Be careful with precedence!
#1 (generic function with 1 method)
julia> (x -> 2*x) ← (x -> x+7) ← Some(3)
Some(20)

Results.:→ — Function

Argument-swapped version of try_map. Enter as \rightarrow.

Example

julia> Some(5) → x -> 2*x
Some(10)
julia> Ok(5) → (x -> 2*x) → (x -> x+7)  # arrows in Julia are right-associative!
ERROR: MethodError
julia> Ok(5) → (x -> 2*x) ∘ (x -> x+7) # but this will work
Ok(24)

Results.:⊗ — Function

Shorthand for and_then (monadic bind). Enter with \otimes

Results.strip_option_type — Method

strip_option_type(ty::Type)::Type

Return the passed type with one layer of Option values stripped.

Results.strip_result_type — Method

strip_result_type(ty::Type)::Type

Return the passed type with one layer of Result values stripped.

Results.try_collect_option — Function

try_collect_option(iter)::Option{Vector}
try_collect_option(options...)::Option{Vector}

Version of try_collect specialized for use with Option Prefer this over try_collect whenever possible to prevent surprising behavior.

Results.try_collect_result — Function

try_collect_result(iter)::Result{Vector}
try_collect_result(results...)::Result{Vector}

Version of try_collect specialized for use with Result. Prefer this over try_collect whenever possible to prevent surprising behavior.

Base methods extended

Base.:! — Method

(!)(result::Result{T,E})::Result{E,T}

Flip an Ok value to an Err value and vice versa.

Base.:& — Method

(&)(result::Result, results...)::Result
(&)(option::Option, options...)::Option

Return the final Ok/Some only if every argument is Ok/Some. Otherwise return the first Err value.

Values can be supplied either as Result/Options or as functions that yield a Result/Option.

Examples

julia> Ok("v1") & Ok("v2") & () -> Ok("v3")
Ok("v3")

julia> Ok("v1") & () -> Err(2) & () -> Ok("v2")
Err(2)

julia> Some(5) & Some(10) & None()

Base.:== — Method

Base.isequal for Err values

Base.:== — Method

Base.isequal for Ok values

Base.:== — Method

Base.isequal for Some values

Base.:| — Method

(|)(result::Result, results...)::Result
(|)(option::Option, options...)::Option

Return the first Ok value found. If no arguments are Ok, return the final Err value.

Values can be supplied either as Results or as functions that yield a Result.

Examples

julia> Err("err1") | Ok("v2") | () -> Ok("v3")
Ok("v2")
julia> Err("err1") | () -> Err("err2") | Err("err3")
Err("err3")

Base.eltype — Method

eltype for Err values

Base.eltype — Method

eltype for None values

Base.eltype — Method

eltype for Ok values

Base.eltype — Method

eltype for Some values

Base.iterate — Method

iterate for Err values

Base.iterate — Method

iterate for None values

Base.iterate — Method

iterate for Ok values

Base.iterate — Method

iterate for Some values

Base.length — Method

length for Err values

Base.length — Method

length for None values

Base.length — Method

length for Ok values

Base.length — Method

length for Some values

Base.show — Method

show for Err values

Base.show — Method

show for Ok values