Results.Types: Core types

Exported members

Results.Types.Err — Type

Represents an computation error.

Results.Types.None — Type

Shorthand for Nothing.

Results.Types.Ok — Type

Represents an Ok result of computation.

Results.Types.Option — Type

Synonym for Union{Some{T}, None}.

Options implement the following protocols:

iterate: Yields one T if the Option is Some, yields nothing otherwise.

length: Returns 1 if the Result is Ok, 0 if Err

Results.Types.Result — Type

Synonym for Union{Ok{T}, Err{E}}.

As well as working with the Result combinators defined below, Results implement the following protocols:

iterate: Yields one T if the Result is Ok, yields nothing otherwise.

length: Returns 1 if the Result is Ok, 0 if Err

Results.Types.none — Constant

Shorthand for nothing.

Un-exported members

Results.Types.UnwrapError — Type

Exception thrown when unwrap() is called on an Err

Base methods extended

Base.:== — Method

Base.isequal for Err values

Base.:== — Method

Base.isequal for Ok values

Base.:== — Method

Base.isequal for Some values

Base.eltype — Method

eltype for Err values

Base.eltype — Method

eltype for None values

Base.eltype — Method

eltype for Ok values

Base.eltype — Method

eltype for Some values

Base.iterate — Method

iterate for Err values

Base.iterate — Method

iterate for None values

Base.iterate — Method

iterate for Ok values

Base.iterate — Method

iterate for Some values

Base.length — Method

length for Err values

Base.length — Method

length for None values

Base.length — Method

length for Ok values

Base.length — Method

length for Some values

Base.show — Method

show for Err values

Base.show — Method

show for Ok values