Results.Collection: Collection utilities

Exported members

Results.Collection.try_first — Method

try_first(s::AbstractString, n::Integer)::Option{AbstractString}

Try to get the first n chars from a string.

Results.Collection.try_first — Method

try_first(c)::Option

Try to get the first element from a collection. Uses Base.firstindex under the hood.

Results.Collection.try_get — Function

Try to get a value from an array or collection.

Results.Collection.try_get — Method

try_get(s::AbstractString, i::Integer)::Option{AbstractChar}

Try to retrieve index index from a string. Uses Base.checkbounds and Base.isvalid under the hood.

Results.Collection.try_get — Method

try_get(collection, index...)::Option

Fallback method for try_get. Relies on exception-handling, so it is slower than the specialized methods.

Results.Collection.try_get — Method

try_get(t::NamedTuple, k::Union{Integer, Symbol})::Option

Try to retrieve member k from a Base.NamedTuple.

Results.Collection.try_get — Method

try_get(a::AbstractArray{T}, index::Integer...)::Option{T}

Try to retrieve index index from an array. Uses Base.isassigned under the hood.

Results.Collection.try_get — Method

try_get(d:AbstractDict{>:K, V}, k::K)::Option{V}

Try to retrive key k from a dictionary.

Results.Collection.try_last — Method

try_last(c)::Option

Try to get the last element from a collection.

Results.Collection.try_peek — Function

try_peek(iter; state=missing)::Option

Try to get the next value from an iterator. If state is not missing, use it in the call to iterate.

Results.Collection.try_pop! — Function

Try to pop a value from a collection.

Results.Collection.try_pop! — Method

try_pop!(iter::Iterators.Stateful)::Option

Try to pop a value from a Stateful iterator.

Results.Collection.try_pop! — Method

try_pop!(a::AbstractArray{T})::Option{T}

Try to pop a value from an array.

Results.Collection.try_pop! — Method

try_pop!(d::AbstractDict{>:K, V}, key::K)::Option{Pair{K, V}}

Try to pop a key-value pair from a dictionary.

Results.Collection.try_pop! — Method

try_pop!(d::AbstractDict{>:K, V}, key::K)::Option{V}

Try to pop the value corresponding to key key from a dictionary.

Un-exported members